Code clean-up - Move `SocketMsg` to appropriate location - Move `GetChatRoomIdAsync` to an appropriate order - Add `BaseApiResponse` to simplify API data object validity checks - Fix repeating code by creating `GetFanslyApiData<T>` - Fix potential error for `_socket.CloseAsync()`
